+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Requires PHP `8.4`
 - `Innmind\OperatingSystem\Remote::sql()` now returns an `Innmind\Immutable\Attempt`
+- `Innmind\OperatingSystem\Filesystem::mount()` now longer automatically create the directory
 
 ## 6.2.0 - 2025-08-10